fix: bug related to deleting db items
Fixed a bug which was stopping the deletion of items in the  db, this was due
to there being many foreign keys in the db. The db logic was changed so
that when any row that contains a foreign key is deleted all rows
containing references to that key are also deleted.